III:15.3: Add special support for modulos and exponentiation operators
Modulo conventionally uses '%' and exponentiation '^'.
I don't intend to add builtin bitwise operators so this should be ok.
Maybe as a library module table.

NOTE: it's incredibly likely that the implementation is inefficient.
We use the C standard math functions `fmod` and `pow` (for doubles).
I'm not terribly concerned with being super duper fast in this project.