III:17.7: Add support for actually compiling expressions
For the most part I follow the chapter word for word.
I do prefer that the giant `rules` lookup table be in its own dedicated
file, as it takes up so much visual space.
This does mean that the various parser functions cannot be static as
they need to be visible to outsiders, hence the use of `extern`.