III:19.5: Add challenge to use flexible array members
I actually quite like this approach because it feels cool. Although it's
restricted to C99 and above, which in the official Lua implementation
would not fly. They very strictly target to standard C89 which is
respectable but I can't see myself doing that, because C89 lacks a lot
of nice features that we take for granted.