III:22.4: Update variable assignment and local scope semantics
Specifically I changed use of `named_variable()` so that it can be
called with an `assignable` boolean. When called by `declaration()` and
a `TOKEN_IDENT` is found, we call it with assignable==true. Otherwise in
expressions we set it to false.

This allows us to forbid users from nesting multiple assignment
expressions and from putting assignment in function calls and the like.

e.g. these are now invalid:

local x=1; local y=2; local z=y=x;
local x=1; print(x=2);

NOTE:

This deviates greatly from Lox as we no longer have the boolean in our
`ParseFn` typedef and also removed it as a param from all prefix/infixfns.