III:24.5.4: Add rudimentary function calls that implicitly return nil
This took quite a bit of work to wrap my head around but in the end it
was so worth it! We'll work on explicitly returning values next. I'm
also interested in replicated the Lua C API `LexState` struct which
seems to combine the Lox structs `Parser` and `Scanner` into one.