Add native functions and replace print statement with a function
TODO: Make the NativeFn signature simpler by just accepting a LVM*,
since in the Lua C API argc is known to be at the top of the stack. With
that information we can deduce where argc is and where the bottom of the
current function should be relative to the VM's stack.