Add local functions, change assignment to be statement, not expression
The changes to assignment semantics allow us to consider the usage of
the assignment operator '=' outside of statments as errors, which is how
Lua does it and how I'd like to do it as well. Treating assignments as
expressions that produce a value can easily be mistaken for checking for
equality, e.g. `local i=0; while (i=4) ... i=i+1 end`.

TODO: Replace Word, DWord and QWord with Byte2, Byte4 and Byte8 structs