Change implementation of Chunk to use Table + TArray for better lookup
We now map constant values to indexes into a values array. Instead of
linear searching the array everytime, we first query the table portion
to see if such a constant was mapped beforehand. If yes, we return the
mapped index. Otherwise we append to the values array and make the new
mapping.