Change C function semantics to allow determining argc at runtime
We now always change the stack window for every function call no matter
what. That is, even C functions cause the VM's bp to be set to point to
the first argument (if there is one). We use the difference between the
VM's sp and the bp to calculate how many arguments are provided for the
function at bp - 1.