15.3: Hacky fix to maintain stack top pointer somewhat properly
This relies on a dangerous assumption that if both arguments B and C are
registers then the stack top pointer needs to be decrement. For most of
our current arithmetic operations this works because OP_ADD will usually
set R(A) with the arguments R(B) + RK(C), where B is usually the same
index as A and C could be a constant index or a register index. My
current handwritten bytecode does push arguments to the stack as needed,
but I imagine this fix won't work well for very long.