15.3: Use bitfields for instructions
Instead of using a simple uint8_t like the book does, this like a more
refined version of the method the Lua C implementation uses. There they
use what would be a `uint32_t` and use a lot of bit manipulation. Here,
we use struct bitfields to achieve a similar goal but with a lot less
manual bit shifting on our part. Our instructions still add up to 32
bits but it's a lot nicer to manage like this in my opinion.