Make concat strings function intern uninterned strings
Because I don't implement a `takeString()` function as in the book, I
completely neglected to check if `concat_strings()` was actually
interning strings if it didn't find them in the VM's interned ones.

I've also changed up the semantics of the `set_*` macros so that they
will not evaluate `v` more than once. This also helps when the value to
be set also evaluates the same thing being pointed to by `v`, such as in
`OP_EQ`. If we set the tagtype BEFORE evaluating `values_equal` in that
case we'd get a completely wrong answer!