Add support for multiple/compound assignment (both global and local)
Later on this will be made more complicated by function calls that may
potentially have 0, 1 or more return values. However we can take
concepts like discarding the expression list if there are too many for
the amount of assignments.

In Lua, multiple return values only work properly when the function call
is the only expression. Otherwise, the function return value/s will be
truncated to 1. If there are no return values then `nil` will be pushed
in its place.