Update multival implementation, make `print` more function-like
It's probably not any more efficient, but I played around with using
function pointers and void pointers as context to "generic" functions.

Of note are `parse_namelist` which receives either `parse_local` or
`parse_variable` as its callback function. For `assign_variables`, we
also pass an array of a new struct called `Assign` which we use to help
resolve multivalue assignment.

It's messy and admittedly probably more work than just repeating
yourself once or twice. But it's fun!