Add scaffolding for tables, start redo of multival implementation
Now considering that it's possible to not only assign a value to an
assignee expression that contains an arbitrary number of `OP_GETTABLE`
instructions, I need to revamp the way I handle multival assignment.

I'm thinking that we can resolve and compile the assignments in order
but keep track of where in the bytecode each assignee expressions starts
at and how many bytes they take up. We do the same for each value
expression after the '='.