Slight refactoring, attempt to allow identifiers in table constructors
I've changed many instances of `Token *` to be just `Token` as it
doesn't seem like `Token` itself is a particularly heavy type and we
never modify the pointer's contents anyway. This will also be useful if
we do recursive functions.

Currently whenever we encounter an identifier in a table constructor we
HAVE to match a '=', indicate we are setting a string key. However, it's
also possible that the user intended the identifier as an element in the
array portion.

TODO: Work on separating hash table from arrays in table implementation