Refactor to allow Lexer to allocate constants directly
Since it carries the VM around we can afford to make this change. This
eases off a little bit of burden from `parser.c` but at the cost of
`lexer.c`. Oh well!

We needed this change to be able to emit both normal strings with escape
sequences and multiline strings with no escape sequences whatsoever.
Instead of creating separate tokens we can now just emit them directly,
so that nobody else throughout the process needs to care about what's a
multiline string and what's not. To the VM, compiler, they're all just
strings.