Revert back to passing all structure parameters as pointers
I gave it some thought and realized for the most part there is not much
copying that needs to be done, so it's generally a better idea to pass
by pointer rather than by value. When we do need to copy we can
explicitly dereference the pointer and let C do automatic shallow
copying for us.